dryrun: don't give up too easily, and add "repeated" annotations
Before, we were giving up when e.g. a description couldn't be eval'ed,
without printing children.
Also, we indicate when a testset is supposed to be repeated, whether
when we don't know how many times (e.g. couldn't eval the length
of testset-for iterator), or when the testset-for description is
a String (doesn't depend on loop variables).