register reliably tested modules
This avoids having to crawl recursively through Base.loaded_modules().
We will still crawl for a little while just to check this works reliably.

Credit to Takafumi Arakaki for the idea of creating a submodule
within modules using InlineTest in order to be able to define
an `__init__` function which does the registration
(as we can't directly create/modify `__init__` for the given module).