don't error when a passed module doesn't have tests
This was the behavior of few commits ago. Now, we don't artificially
add empty tests, we rather simply filter modules out.
This can happen legitly when a parent module is passed with the
idea to test its submodules.

In the future, we could error when a module is passed and doesn't have
tests even recursively.

We still error when the final list of modules is empty.