don't always print module header when only one testset
When only one module was tested and explicitly passed to `retest`,
`moduleheader` was false, but we still printed the header when
there was only one toplevel testset, so that the module name was
printed at least once. But this was not very consistent: depending
on the filtering pattern, the module name could be printed either
at the beginning or at the end. When the module is passed explicitly,
it doesn't seem necessary to print the module name at all, so we only
print it at the end when a summary is needed, i.e. more than one
toplevel testset was tested.