print module header when it doesn't match the passed module
We don't print the module header when only one module is tested, and
this module is passed explicitly as an argument to `retest`. But we
were wrongly not printing this header when one module not containing
tests was passed, which was containing a submodule with tests, and
`recursive=true`. In this case, we want to print the submodule header,
as this submodule was no passed explicitly to `retest`
(idem with `MyPackageTests` and `load=true`).